// terraform/application/types.d.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType } from 'kysely';

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export interface TerraformState {
  content: string;
  created_at: Generated<number>;
  id: Generated<string>;
  name: string;
  updated_at: Generated<number>;
}

export interface DB {
  terraform_state: TerraformState;
}